14 fix

+++ 14-binary_tree_balance.c
@@ -7,9 +7,9 @@
  *
  * Return: Height of the tree. If tree is NULL, return 0.
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
-        int lh, rh;
+        size_t lh, rh;
 
         if (tree == NULL)
                 return (0);
